coreos/requirements.txt
jinja2
pyYAML
